refactor: relocate ChannelOverviewProvider to agent package
ChannelOverviewProvider assembles a compact JSON summary of active
communication channels for system-prompt injection. It is a composition,
not a domain primitive: it consumes the loop registry, the contact store,
and request hints to describe the agent's current communication
landscape.

That makes it agent-level work. It already co-locates well there —
agent/ already owns ChannelProvider, which the composite adds alongside
the overview in app/new_awareness.go. Splitting the two across packages
was an accident of history.

Move the provider, its interfaces (ChannelLoopSource, PhoneResolver,
HintsFunc, LoopSnapshot), and its test into internal/agent/. Update the
two wiring sites in app/new_awareness.go and app/adapters.go. The
awareness import in adapters.go was load-bearing only for the old
interfaces, so it drops out entirely.

With this move plus the earlier timefmt extraction and StateWindow
relocation, #625's definition of success is met: no relocated
context provider imports internal/awareness any longer.

Closes #625.
