+++ internal/agent/channel_overview_test.go
@@ -1,4 +1,4 @@
-package awareness
+package agent
 
 import (
 	"context"
